added Home.js as a components
To explain/ask:

First there is the gray navbar, which I can turn into a reusable component if need be, and the home icon is a button which can redirect wherever needed. The the title which is welcome student, it will use dynamic names depending on who is logged in, so i need backend for that. The useState hook is temporary because idk how the backend looks. then the li elements which are classes are buttons which i assume will redirect somewhere as well, I can change them to links if needed. So far theyre hardcoded in to showcase but technically they will also change based on backend functionality.

Ask what backend framework is being used. Idk what else, all of this front end is hardcoded just to show, but has no functionality.